backend/requirements.txt
fastapi
sqlalchemy
apscheduler
praw
spacy
nltk
requests
python-dotenv
psycopg2-binary
